+++ examples/react-app/src/App.tsx
@@ -1,12 +1,13 @@
 import "./App.css";
 import { Web3Auth } from "@web3auth/web3auth";
-import { ADAPTER_EVENTS, CHAIN_NAMESPACES } from "@web3auth/base";
+import { ADAPTER_EVENTS } from "@web3auth/base";
 import { LOGIN_MODAL_EVENTS } from "@web3auth/ui";
 import { useEffect, useState } from "react";
+import { useWeb3Auth } from "./components/Web3AuthProvider";
 
 function App() {
   const [user, setUser] = useState(null);
-  const [web3auth, setWeb3auth] = useState<Web3Auth | null>(null);
+  const web3auth = useWeb3Auth();
   const [loaded, setLoaded] = useState(false);
 
   useEffect(() => {
@@ -36,23 +37,6 @@ function App() {
       });
     };
 
-    const polygonMumbaiConfig = {
-      chainNamespace: CHAIN_NAMESPACES.EIP155,
-      rpcTarget: "https://rpc-mumbai.maticvigil.com",
-      blockExplorer: "https://mumbai-explorer.matic.today",
-      chainId: "0x13881",
-      displayName: "Polygon Mumbai Testnet",
-      ticker: "matic",
-      tickerName: "matic",
-    };
-
-    const web3auth = new Web3Auth({
-      chainConfig: polygonMumbaiConfig,
-      clientId: "BKPxkCtfC9gZ5dj-eg-W6yb5Xfr3XkxHuGZl2o2Bn8gKQ7UYike9Dh6c-_LaXlUN77x0cBoPwcSx-IVm0llVsLA",
-    });
-
-    setWeb3auth(web3auth);
-
     // ⭐️ initialize modal on page mount.
     const initializeModal = async () => {
       console.log("initializeModal");
@@ -65,16 +49,14 @@ function App() {
   }, []);
 
   const login = async () => {
-    if (!web3auth) return;
     const provider = await web3auth.connect();
     // TODO: add this provider to web3/ethers
   };
   const logout = async () => {
-    if (!web3auth) return;
     await web3auth.logout();
   };
+
   const getUserInfo = async () => {
-    if (!web3auth) return;
     const userInfo = await web3auth.getUserInfo();
     console.log(userInfo);
   };
@@ -102,7 +84,7 @@ function App() {
     );
   };
 
-  return loaded ? (user ? renderAuthenticated() : renderUnauthenticated()): <h1>Loading....</h1>;
+  return loaded ? user ? renderAuthenticated() : renderUnauthenticated() : <h1>Loading....</h1>;
 }
 
 export default App;
